+++ e-Homegreen/ScanEventsViewController.swift
@@ -375,38 +375,20 @@ class ScanEventsViewController: PopoverVC, ProgressBarDelegate {
         fromTextField.text = ""
         toTextField.text = ""
     }
-
+    
     @IBAction func btnRemove(sender: UIButton) {
-        
-        let optionMenu = UIAlertController(title: nil, message: "Are you sure you want to delete all scenes?", preferredStyle: .ActionSheet)
-        let deleteAction = UIAlertAction(title: "Delete", style: .Default, handler: {
-            (alert: UIAlertAction!) -> Void in
-            
-            if self.events.count != 0 {
-                for event in self.events {
-                    self.appDel.managedObjectContext!.deleteObject(event)
+        showAlertView(sender, message: "Are you sure you want to delete all scenes?") { (action) in
+            if action == ReturnedValueFromAlertView.Delete{
+                if self.events.count != 0 {
+                    for event in self.events {
+                        self.appDel.managedObjectContext!.deleteObject(event)
+                    }
                 }
-                
+                CoreDataController.shahredInstance.saveChanges()
+                self.refreshEventList()
+                self.view.endEditing(true)
             }
-            CoreDataController.shahredInstance.saveChanges()
-            self.refreshEventList()
-            self.view.endEditing(true)
-        })
-        
-        let cancelAction = UIAlertAction(title: "Cancel", style: .Cancel, handler: {
-            (alert: UIAlertAction!) -> Void in
-            print("Cancelled")
-        })
-        
-        if let popoverController = optionMenu.popoverPresentationController {
-            popoverController.sourceView = sender
-            popoverController.sourceRect = sender.bounds
         }
-        
-        optionMenu.addAction(deleteAction)
-        optionMenu.addAction(cancelAction)
-        self.presentViewController(optionMenu, animated: true, completion: nil)
-
     }
     
     // MARK: - FINDING EVENTS
